Refactor AVD Manager
