+++ virtualMachineManager.go
@@ -3,7 +3,6 @@ package cloudyazure
 import (
 	"context"
 	"fmt"
-	"time"
 
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore/arm"
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore/cloud"
@@ -34,8 +33,6 @@ type AzureVirtualMachineManager struct {
 	usageClient *armcompute.UsageClient
 
 	galleryClient *armcompute.SharedGalleryImageVersionsClient
-
-	avdManager *AzureVirtualDesktopManager
 
 	LogBody bool
 }
@@ -113,12 +110,6 @@ func (vmm *AzureVirtualMachineManager) Configure(ctx context.Context) error {
 	}
 	vmm.usageClient = usageClient
 
-	avdManager, err := NewAzureVirtualDesktopManager(ctx, vmm.credentials, &vmm.config.AvdConfig)  // TODO: use AVD only creds
-	if err != nil {
-		return err
-	}
-	vmm.avdManager = avdManager
-
 	return nil
 }
 
@@ -185,163 +176,12 @@ func (vmm *AzureVirtualMachineManager) Update(ctx context.Context, vm *models.Vi
 	return nil, nil
 }
 
-func (vmm *AzureVirtualMachineManager) AvdRegister(ctx context.Context, vm *models.VirtualMachine, domainName, domainUsername, domainPassword string) (*models.VirtualMachine, error) {
-	log := logging.GetLogger(ctx)
-	
-	log.InfoContext(ctx, "Configuring AVD for VM", "VM", vm.ID)
+func UpdateCloudyVirtualMachine(vm *models.VirtualMachine, responseVirtualMachine armcompute.VirtualMachine) error {
 
-	var err error
-	
-	start := time.Now()
-	hostPool, regToken, err := vmm.avdManager.PreRegister(ctx, vm, avdRgName)
-	log.DebugContext(ctx, "AVD Pre-Register completed", "time", time.Since(start).String())
-	if err != nil {
-		return nil, errors.Wrap(err, "AVD Pre-register")
-	}
-
-	start = time.Now()
-	err = vmm.runAvdRegistrationScript(ctx, vm, *regToken, domainName, domainUsername, domainPassword)
-	log.DebugContext(ctx, "AVD Registration Script completed in %s", "time", time.Since(start).String())
-	if err != nil {
-		return nil, errors.Wrap(err, "AVD Register Script")
-	}
-
-	start = time.Now()
-	connection, err := vmm.avdManager.PostRegister(ctx, avdRgName, *hostPool.Name, vm.UserID, vm)
-	log.DebugContext(ctx, "AVD Post-Register completed in %s", "time", time.Since(start).String())
-	if err != nil {
-		return nil, errors.Wrap(err, "AVD Post-register")
-	}
-
-	vm.Connect = connection
-	return vm, nil
+	return nil
 }
 
-func (vmm *AzureVirtualMachineManager) runAvdRegistrationScript(ctx context.Context, vm *models.VirtualMachine, registrationKey, domainName, domainUsername, domainPassword string) error {	
-	rgName := vmm.credentials.ResourceGroup
-	vmName := vm.ID
-	osType := vm.Template.OperatingSystem
-
-	if osType != "windows" {
-		return errors.New("unsupported OS type; only Windows is supported for AVD registration")
-	}
-
-	
-	// Define the PowerShell script with placeholders
-	scriptTemplate := `
-# Set up logging for verbose output
-$logFilePath = "$pwd\install_log.txt"
-Start-Transcript -Path $logFilePath -Append
-
-$REGISTRATIONTOKEN = "%s"
-$DOMAIN_NAME = "%s"
-$DOMAIN_USERNAME = "%s"
-$DOMAIN_PASSWORD = "%s"
-
-# Check if the machine is already in the domain
-$computerDomain = (Get-WmiObject -Class Win32_ComputerSystem).Domain
-if ($computerDomain -eq $DOMAIN_NAME) {
-    Write-Host "Machine is already part of the domain: $DOMAIN_NAME"
-} else {
-    # Join the domain
-    try {
-        Write-Host "Attempting to join the domain: $DOMAIN_NAME"
-        $securePassword = ConvertTo-SecureString -String $DOMAIN_PASSWORD -AsPlainText -Force
-        $credential = New-Object System.Management.Automation.PSCredential ($DOMAIN_USERNAME, $securePassword)
-        Add-Computer -DomainName $DOMAIN_NAME -Credential $credential -Force -Verbose
-        Write-Host "Successfully joined the domain."
-    } catch {
-        Write-Host "Error joining the domain: $_"
-        Stop-Transcript
-        return
-    }
-}
-
-# Define URLs for the installers
-$uris = @(
-	"https://query.prod.cms.rt.microsoft.com/cms/api/am/binary/RWrmXv",   # RDAgent
-	"https://query.prod.cms.rt.microsoft.com/cms/api/am/binary/RWrxrH"    # BootLoader Agent
-)
-
-$installers = @()
-
-# Download installers
-foreach ($uri in $uris) {
-	try {
-		Write-Host "Starting download: $uri"
-		$download = Invoke-WebRequest -Uri $uri -UseBasicParsing -Verbose
-		$fileName = ($download.Headers.'Content-Disposition').Split('=')[1].Replace('"','')
-		$outputPath = "$pwd\$fileName"
-
-		if (Test-Path $outputPath) {
-			Write-Host "File $fileName already exists. Skipping download."
-		} else {
-			$output = [System.IO.FileStream]::new($outputPath, [System.IO.FileMode]::Create)
-			$output.write($download.Content, 0, $download.RawContentLength)
-			$output.close()
-			Write-Host "Downloaded: $fileName"
-		}
-
-		$installers += $outputPath
-	} catch {
-		Write-Host "Error downloading ${uri}: $_"
-		return
-	}
-}
-
-# Unblock files after download
-foreach ($installer in $installers) {
-	if (Test-Path $installer) {
-		Write-Host "Unblocking file: $installer"
-		Unblock-File -Path $installer -Verbose
-	} else {
-		Write-Host "File $installer not found, skipping unblock."
-	}
-}
-
-# Find the RDAgent installer
-$rdaAgentInstaller = $installers | Where-Object { $_ -match "Microsoft.RDInfra.RDAgent.Installer-x64" }
-if (-not $rdaAgentInstaller) {
-	Write-Host "RDAgent installer not found."
-	return
-}
-
-# Find the BootLoader installer
-$rdaBootLoaderInstaller = $installers | Where-Object { $_ -match "Microsoft.RDInfra.RDAgentBootLoader.Installer-x64" }
-if (-not $rdaBootLoaderInstaller) {
-	Write-Host "BootLoader Agent installer not found."
-	return
-}
-
-# Install RDAgent
-Write-Host "Installing RDAgent with registration token."
-try {
-	Start-Process msiexec -ArgumentList "/i $rdaAgentInstaller REGISTRATIONTOKEN=$REGISTRATIONTOKEN /quiet /norestart" -Wait -Verbose -Verb RunAs
-} catch {
-	Write-Host "Error installing RDAgent: $_"
-	return
-}
-
-# Install BootLoader Agent
-Write-Host "Installing BootLoader Agent."
-try {
-	Start-Process msiexec -ArgumentList "/i $rdaBootLoaderInstaller /quiet" -Wait -Verbose -Verb RunAs
-} catch {
-	Write-Host "Error installing BootLoader Agent: $_"
-	return
-}
-
-# Finalize and restart
-Write-Host "Preparing for system restart."
-Restart-Computer -Force
-
-# Stop the transcript and finalize the log
-Stop-Transcript
-	`
-
-	// Inject the registration key and domain credentials into the script
-	script := fmt.Sprintf(scriptTemplate, registrationKey, domainName, domainUsername, domainPassword)
-
+func (vmm *AzureVirtualMachineManager) RunPowershell(ctx context.Context, vmID, script string) error {
 	// Define RunCommandInput
 	runCommandInput := armcompute.RunCommandInput{
 		CommandID: to.Ptr("RunPowerShellScript"),
@@ -351,9 +191,9 @@ Stop-Transcript
 	}
 
 	// Execute the script
-	response, err := vmm.vmClient.BeginRunCommand(ctx, rgName, vmName, runCommandInput, nil)
+	response, err := vmm.vmClient.BeginRunCommand(ctx, vmm.credentials.ResourceGroup, vmID, runCommandInput, nil)
 	if err != nil {
-		return errors.Wrap(err, "failed to execute AVD registration script")
+		return errors.Wrap(err, "failed to execute remote powershell script")
 	}
 
 	// Poll until the command completes
@@ -372,11 +212,6 @@ Stop-Transcript
 	} else {
 		fmt.Println("No output returned from the command.")
 	}
-
-	return nil
-}
-
-func UpdateCloudyVirtualMachine(vm *models.VirtualMachine, responseVirtualMachine armcompute.VirtualMachine) error {
 
 	return nil
 }
