+++ virtualMachineManagerUtilities.go
